microservices/src/project_files/requirements.txt
fastapi==0.83.0
uvicorn==0.16.0
pymongo[srv]
PyJWT
python-jose
mangum